# src/main/resources/openApi/api.yaml
openapi: "3.0.0"
info:
  title: Api for credit card print
  description: |
    The API enables printing of credit cards for individuals.  
    In the first step, the data for printing is sent.  
    In the next step, we send a request for printing.  
    It is advisable to check whether the printing was successful after a certain time period (1-2 minutes) by using the search endpoint.  
    Other functionalities are searching and deletion of credit card print.
  version: 1.0.2
paths:
  /credit-card-print:
    get:
      tags:
        - search
      summary: 'Search for credit card'
      description: |
        You can perform a search using one of the optional parameters: first name, last name, Personal Identification Number (OIB), or status.  
        If a parameter is not specified, it will return all results.    
        If no matching records are found, an empty list will be returned.
      operationId: searchCreditCard
      parameters:
          - in: query
            name: firstName
            schema:
             type: string
            required: false
            description: Search by first name
            example: Danica
          - in: query
            name: lastName
            schema:
             type: string
            required: false
            description: Search by last name
            example: Marković
          - in: query
            name: oib
            schema:
             type: string
            required: false
            description: Search by user personal identification number (OIB)
            example: 42585672100
          - in: query
            name: status
            schema:
              type: string
            required: false
            description: Search by status
            example: NO_ACTIVE
      responses:
        '2XX':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                 $ref: '#/components/schemas/CreditCardPrintDetailsDto'
                example:
                - firstName: Josip
                  lastName: Pavić
                  oib: 84574720654
                - firstName: Aron
                  lastName: Ivanković
                  oib: 31730672508
        '4XX':
          description: The server cannot or will not process the request due to malformed request.
        '5XX':
          description: Problem on the server, try again latter.
        default:
          description: Unexpected error
    post:
      tags:
        - print request
      summary: Add a new credit card for print
      description: |
        The data for credit card print is sent to this endpoint.  
        Sets the status "NO_ACTIVE," which indicates that the data has been successfully entered into the system.    
        Returns created entity.
      operationId: insertCreditCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditCardPrintInsertDto'
            examples:
              Example 1:
                value:
                  firstName: Sandi
                  lastName: Vuković
                  oib: 32052174042
              Example 2:
                value:
                  firstName: Matej
                  lastName: Župan
                  oib: 06675310266
      responses:
        '2XX':
          description: OK
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CreditCardPrintDetailsDto'
              examples:
                Example 1:
                  value:
                    firstName: Paško
                    lastName: Horvatinčić
                    oib: 94145125477
                    status: IN_PROGRESS
        '4XX':
          description: Client error
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/ValidationErrorResponse'
              example:
                - {'fieldName': 'oib', 'message': 'Invalid OIB'}
        '5XX':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/ErrorResponse'
              example:
                - status : CONFLICT
                  message : "Credit Card Print {oib: 98614461866} already exist and it is in status BLOCKED"
        default:
          description: Unexpected error
    put:
      tags:
        - print
      summary: For individual found by oib generate a text file a start process of credit card printing.
      description: |
        Process of credit card printing can be started only for individual person credit card print request in status NO_ACTIVE.  
        Returns the status indicating success of process.  
        Returns IN_PROCESS when file is generated. Printing process is started, waiting for result.  
        Returns ACTIVE if credit card is printed.  
        Returns BLOCKED if credit card is not printed. Process of printing failed.
      operationId: printCreditCard
      parameters:
        - in: query
          name: oib
          required: true
          schema:
            type: string
          example: 68695906587
      responses:
        '2XX':
          description: OK
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CreditCardPrintStatusDto'
              examples:
                Example 1:
                  value:
                    oib: 22782713620
                    processStatus: REQUEST_SUCCESSFULLY_WAITING_PRINTER
                    msg: null
                Example 2:
                  value:
                    oib: 22782713620
                    processStatus: FAILURE
                    msg: 'For {oib: 22782713620} Credit Card Print does not exist'
        '4XX':
          description: Client error
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/ValidationErrorResponse'
              example:
               - { 'fieldName': 'oib', 'message': 'Invalid OIB' }
        '5XX':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/ErrorResponse'
              example:
                - status: INTERNAL_SERVER_ERROR
                  message: "Exception in toString()"
        default:
          description: Unexpected error
    delete:
      tags:
        - delete
      summary: Delete credit card print record
      description: |
        Record in database is deleted and.  
        Add _DELETED postfix to file name (if has any).
      operationId: deleteCreditCard
      parameters:
        - in: query
          name: oib
          required: true
          schema:
            type: string
          example: 76238827705
      responses:
        '2XX':
          description: OK - deleted
        default:
          description: Unexpected error
components:
  schemas:
    CreditCardPrintDetailsDto:
      type: object
      required:
        - firstName
        - lastName
        - oib
        - status
      properties:
        firstName:
          type: string
        lastName:
          type: string
        oib:
          type: string
        status:
          type: string
    CreditCardPrintInsertDto:
      type: object
      required:
        - firstName
        - lastName
        - oib
      properties:
        firstName:
          type: string
        lastName:
          type: string
        oib:
          type: string
          x-validations-oib: true
    CreditCardPrintStatusDto:
      type: object
      required:
        - oib
        - processStatus
      properties:
        oib:
          type: string
        processStatus:
          type: string
          enum: [REQUEST_SUCCESSFULLY_WAITING_PRINTER, FAILURE]
        msg:
          type: string
    ValidationErrorResponse:
      type: object
      properties:
        violations:
          type: array
          items:
           $ref: '#/components/schemas/Violation'
          description: List of the violations
    Violation:
      type: object
      properties:
        fieldName:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string